Add new class files for a-codequality increment
